+++ webapp/channels/src/types/store/draft.ts
@@ -3,6 +3,7 @@
 
 import type {FileInfo} from '@mattermost/types/files';
 import type {PostPriority} from '@mattermost/types/posts';
+import type {ScheduledPost} from '@mattermost/types/schedule_post';
 
 export type DraftInfo = {
     id: string;
@@ -28,3 +29,19 @@ export type PostDraft = {
         };
     };
 };
+
+export function scheduledPostToPostDraft(scheduledPost: ScheduledPost): PostDraft {
+    return {
+        message: scheduledPost.message,
+        fileInfos: scheduledPost.metadata?.files || [],
+        uploadsInProgress: [],
+        props: scheduledPost.props,
+        channelId: scheduledPost.channel_id,
+        rootId: scheduledPost.root_id,
+        createAt: 0,
+        updateAt: 0,
+        metadata: {
+            priority: scheduledPost.priority,
+        },
+    };
+}
